Add support for Windows with code from pemjax
The code comes from project pemjax with minor modification
by xionghuaidong.

The pemjax code comes from:
  https://pypi.tuna.tsinghua.edu.cn/simple/pemjax/

The following pemjax repositories is not available:
  https://github.com/IGinX-THU/pemja
  https://github.com/IGinX-THU/pemjax

+++ src/main/c/pemja/core/pylib.c
@@ -602,7 +602,7 @@ JcpPyObject_SetJLong(JNIEnv *env, intptr_t ptr, const char *name, jlong value)
 
     Jcp_BEGIN_ALLOW_THREADS
 
-    _JcpPyObject_SetPyObject(jcp_thread->globals, name, JcpPyInt_FromLong((long) value));
+    _JcpPyObject_SetPyObject(jcp_thread->globals, name, JcpPyInt_FromLong(value));
 
     Jcp_END_ALLOW_THREADS
 }
